Update README.md
Title : Autonomous Toolkit to Forecast Customers Churn
Description
The aim of this project is to automate the identification of customers at risk of churning by using Data
Science techniques. The product enables business managers to make appropriate decision to retain customers
at risk. The toolkit shall be performed with user’s consent showing them iteratively the summary and the impact
of the changes they might make on the performance of the technique or the prediction.